+++ web/app/themes/rebe/src/assets/scripts/workshops/directives/preview.directive.js
@@ -1,0 +1,12 @@
+angular
+  .module('workshops.preview.directive', ['ui.router'])
+  .directive('workshopPreview', function () {
+    return {
+      restrict: 'E',
+      replace: true,
+      templateUrl:  '/app/themes/rebe/templates/content/post-types/workshop/ng/preview.html',
+      link: function (scope, element) {
+        //
+      }
+    };
+  });