+++ web/app/themes/rebe/templates/layouts/base-404.php
@@ -2,7 +2,7 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <?php if (is_page_template('templates/template-events.php') || is_page_template('templates/template-stories.php')) : ?>
+  <?php if (is_page_template('templates/template-events.php') || is_page_template('templates/template-stories.php') || is_page_template('templates/template-workshops.php')) : ?>
   <?php App\set_template_uri_base(); ?>
   <?php endif; ?>
   <?php wp_head(); ?>